# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rotpunkt Translate",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/projects": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create Project",
        "description": "Upload a PDF, parse its contents, detect languages, and create a new project.\n\nArgs:\n    file: The uploaded PDF file payload.\n    db_session: The active asynchronous database session.\n\nReturns:\n    ProjectResponse containing the new project's metadata, parsed pages, and segments.\n\nRaises:\n    HTTPException: 400 for missing filename or bad PDF, 500 for unhandled parsing errors.",
        "operationId": "create_project_api_projects_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_project_api_projects_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Project",
        "description": "Get a project and all its parsed contents.\n\nArgs:\n    project_id: The UUID of the project to retrieve.\n    db_session: The active asynchronous database session.\n\nReturns:\n    ProjectResponse representing the requested project.",
        "operationId": "get_project_api_projects__project_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "Update Project",
        "description": "Update a project, typically to confirm or change its source language.\n\nChanging the source language recalculates the 'is_translatable' flag\nfor every segment in the project.\n\nArgs:\n    project_id: The UUID of the project to update.\n    payload: ProjectUpdateRequest containing the new source language.\n    db_session: The active asynchronous database session.\n\nReturns:\n    ProjectResponse representing the updated project.",
        "operationId": "update_project_api_projects__project_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/segments/{segment_id}": {
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "Edit Segment",
        "description": "Manually edit the translation for a single document segment.\n\nArgs:\n    project_id: The UUID of the project.\n    segment_id: The UUID of the segment to edit.\n    payload: SegmentEditRequest containing target language and new text.\n    db_session: The active asynchronous database session.\n    current_user: The authenticated user.\n\nReturns:\n    The updated DocumentSegment object.\n\nRaises:\n    HTTPException: 404 if the project or segment is not found.",
        "operationId": "edit_segment_api_projects__project_id__segments__segment_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "segment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Segment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SegmentEditRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentSegmentSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/segments/{segment_id}/retranslate": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Retranslate Segment",
        "description": "Re-translate a single document segment synchronously.\n\nArgs:\n    project_id: The UUID of the project.\n    segment_id: The UUID of the segment to re-translate.\n    payload: TranslateRequest containing the target language code.\n    db_session: The active asynchronous database session.\n\nReturns:\n    The updated DocumentSegment object.\n\nRaises:\n    HTTPException: 404 if the project or segment is not found.\n    HTTPException: 400 if the document has no source language or segment is not translatable.",
        "operationId": "retranslate_segment_api_projects__project_id__segments__segment_id__retranslate_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "segment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Segment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranslateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentSegmentSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/translate": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Translate Project",
        "description": "Trigger the translation of a project.\n\nThis endpoint returns immediately (202 Accepted) while the translation\norchestrator processes the document segments in the background.\n\nArgs:\n    project_id: The UUID of the project to translate.\n    payload: TranslateRequest containing the target language code.\n    background_tasks: FastAPI BackgroundTasks runner.\n    db_session: The active asynchronous database session.\n\nReturns:\n    A dictionary containing a success message indicating background processing started.\n\nRaises:\n    HTTPException: 400 if the project has no documents or is missing a source language.",
        "operationId": "translate_project_api_projects__project_id__translate_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranslateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Translate Project Api Projects  Project Id  Translate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_create_project_api_projects_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_create_project_api_projects_post"
      },
...
      "DocumentSchema": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "source_language": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "de",
                  "en",
                  "fr",
                  "nl",
                  "da",
                  "nb",
                  "es"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Language"
          },
          "pages": {
            "items": {
              "$ref": "#/components/schemas/DocumentPageSchema"
            },
            "type": "array",
            "title": "Pages"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/DocumentSegmentSchema"
            },
            "type": "array",
            "title": "Segments"
          }
        },
        "type": "object",
        "required": [
          "id",
          "original_filename",
          "source_language",
          "pages",
          "segments"
        ],
        "title": "DocumentSchema"
      },
      "DocumentSegmentSchema": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "page_number": {
            "type": "integer",
            "title": "Page Number"
          },
          "order_index": {
            "type": "integer",
            "title": "Order Index"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "bounding_box": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Bounding Box"
          },
          "detected_language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected Language"
          },
          "is_translatable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Translatable"
          },
          "translations": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Translations"
          },
          "annotations": {
            "additionalProperties": {
              "items": {
                "additionalProperties": true,
                "type": "object"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Annotations"
          },
          "issues": {
            "additionalProperties": {
              "items": {
                "additionalProperties": true,
                "type": "object"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "id",
          "page_number",
          "order_index",
          "kind",
          "text",
          "bounding_box",
          "detected_language",
          "is_translatable",
          "translations"
        ],
        "title": "DocumentSegmentSchema"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentSchema"
            },
            "type": "array",
            "title": "Documents"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "created_at",
          "updated_at",
          "documents"
        ],
        "title": "ProjectResponse"
      },
      "ProjectStatus": {
        "type": "string",
        "enum": [
          "new",
          "translating",
          "review",
          "done"
        ],
        "title": "ProjectStatus"
      },
      "ProjectUpdateRequest": {
        "properties": {
          "source_language": {
            "type": "string",
            "enum": [
              "de",
              "en",
              "fr",
              "nl",
              "da",
              "nb",
              "es"
            ],
            "title": "Source Language"
          }
        },
        "type": "object",
        "required": [
          "source_language"
        ],
        "title": "ProjectUpdateRequest"
      },
...
      "SegmentEditRequest": {
        "properties": {
          "target_language": {
            "type": "string",
            "enum": [
              "de",
              "en",
              "fr",
              "nl",
              "da",
              "nb",
              "es"
            ],
            "title": "Target Language"
          },
          "new_text": {
            "type": "string",
            "title": "New Text"
          }
        },
        "type": "object",
        "required": [
          "target_language",
          "new_text"
        ],
        "title": "SegmentEditRequest"
      },
...
      "TranslateRequest": {
        "properties": {
          "target_language": {
            "type": "string",
            "enum": [
              "de",
              "en",
              "fr",
              "nl",
              "da",
              "nb",
              "es"
            ],
            "title": "Target Language"
          }
        },
        "type": "object",
        "required": [
          "target_language"
        ],
        "title": "TranslateRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/auth/token"
          }
        }
      }
    }
  }
}